api/requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
black==21.7b0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
Flask-Caching==1.10.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.0
pathspec==0.9.0
Pillow==8.3.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
python-telegram-bot==13.7
pytz==2021.1
PyWavelets==1.1.1
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
scipy==1.7.0
six==1.16.0
tomli==1.2.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
git+https://github.com/ozgur/python-firebase
opencv-python-headless
flask-swagger-ui==3.20.9
